requirements.txt
colorama==0.4.4
commonmark==0.9.1
cursor==1.3.4
pychalk==2.0.1
pydub==0.25.1
Pygments==2.8.1
six==1.15.0
terminal-animation==0.6
typing-extensions==3.7.4.3
